moved topology to top level
