+++ pygeo/__init__.py
@@ -6,6 +6,7 @@ from .pyBlock import pyBlock
 from .constraints.DVCon import DVConstraints
 from .parameterization.DVGeo import DVGeometry
 from .parameterization.DVGeometryAxi import DVGeometryAxi
+from .topology import Topology
 
 try:
     from .parameterization.DVGeometryVSP import DVGeometryVSP
