+++ pygeo/geo_utils/__init__.py
@@ -19,7 +19,6 @@ from .projection import *  # noqa: F401, F403
 from .remove_duplicates import *  # noqa: F401, F403
 from .rotation import *  # noqa: F401, F403
 from .split_quad import *  # noqa: F401, F403
-from .topology import *  # noqa: F401, F403
 
 
 # Set a (MUCH) larger recursion limit. For meshes with extremely large
